==> ETL_notebooks/part2_omdb_api.ipynb <==
#Dependencies
import pandas as pd
import requests
import json
import time

#API key
from api_keys import omdb_api_key_two

#Read wikipedia csv into DataFrame
all_movies_df = pd.read_csv("../resources/wikipedia_all_movies_2018_to_2022.csv")

#List of wikipedia titles
title_list = [title for title in all_movies_df["Title"]]

#Split title list into two parts
title_list_1 = title_list[:1000]
title_list_2 = title_list[999:]


#List of imdbID's
imdb_ids = []

#Counter
record = 0

#API calls
for title in title_list_2:
    
    try:
        record += 1
        print(f"Movie {record} of {len(title_list_2)}: {title}")
        
        url = f"http://www.omdbapi.com/?t={title}&apikey={omdb_api_key_two}"
        movie_json = requests.get(url).json()
        
        imdb_ids.append({"imdbID":movie_json["imdbID"]})
        
        print("Success")
        time.sleep(1)
        
    except Exception as e:
        print(e)
        pass

#Export to csv
imdb_ids_df = pd.DataFrame(imdb_ids)
imdb_ids_df.to_csv("../output_data/imdb_ids_02_keep.csv",index=False)
