==> ETL_notebooks/part3_imdb_scrape.ipynb <==
#Dependencies
import pandas as pd
import requests
import json
import time
import bs4 as BeautifulSoup

#Read the imdbID lists
list_01 = pd.read_csv("../output_data/imdb_ids_01_keep.csv")
list_02 = pd.read_csv("../output_data/imdb_ids_02_keep.csv")

#Combine the lists
imdb_ids = pd.concat([list_01,list_02])
imdb_ids["imdbID"].count()

#Drop duplicates
imdb_ids_no_dupes = imdb_ids.drop_duplicates(subset=["imdbID"])
imdb_ids_no_dupes["imdbID"].count()

#Get data from IMDb
HEADERS = {"User-Agent": "Chrome/111.0 (PC; Windows 10 Home)"}

movie_data = []

movieid_list = [movieid for movieid in imdb_ids_no_dupes["imdbID"]]

record = 0

for movieid in movieid_list:
    
    try:
        
        url = f"https://www.imdb.com/title/{movieid}/"
        response = requests.get(url, headers=HEADERS)
        soup = BeautifulSoup.BeautifulSoup(response.text,"html.parser")
        money = soup.find("script", id="__NEXT_DATA__")
        json_object = json.loads(money.text)
        
        record += 1
        
        title = json_object["props"]["pageProps"]["aboveTheFoldData"]["titleText"]["text"]
        
        print(f"Movie: {record} of {len(movieid_list)}: {title}")
        
        try:
            mtype = json_object["props"]["pageProps"]["aboveTheFoldData"]["titleType"]["text"]
        except:
            mtype = ""
            print(f"No type")
        
        try:
            imdbbudget = json_object["props"]["pageProps"]["mainColumnData"]["productionBudget"]["budget"]["amount"]
        except:
            imdbbudget = ""
            print(f"No budget")
        
        try:
            wwgross = json_object["props"]["pageProps"]["mainColumnData"]["worldwideGross"]["total"]["amount"]
        except:
            wwgross = ""
            print(f"No gross")
            
        try:
            year = json_object["props"]["pageProps"]["aboveTheFoldData"]["releaseDate"]["year"]
            month = json_object["props"]["pageProps"]["aboveTheFoldData"]["releaseDate"]["month"]
            day = json_object["props"]["pageProps"]["aboveTheFoldData"]["releaseDate"]["day"]
            release = f"{year}-{month}-{day}"
        except:
            release = ""
            print(f"No date")
        
        try:
            runtime = json_object["props"]["pageProps"]["aboveTheFoldData"]["runtime"]["seconds"]
        except:
            runtime = ""
            print(f"No runtime")
        
        try:
            genre = json_object["props"]["pageProps"]["aboveTheFoldData"]["genres"]["genres"][0]["text"]
        except:
            genre = ""
            print(f"No genre")
        
        try:
            noms = json_object["props"]["pageProps"]["mainColumnData"]["prestigiousAwardSummary"]["nominations"]
        except:
            noms = ""
            print(f"No noms")
            
        try:
            mscore = json_object["props"]["pageProps"]["aboveTheFoldData"]["metacritic"]["metascore"]["score"]
        except:
            mscore = ""
            print(f"No mscore")
        
        try:
            imdbRating = json_object["props"]["pageProps"]["aboveTheFoldData"]["ratingsSummary"]["aggregateRating"]
        except:
            imdbRating = ""
            print(f"No rating")
        
        try:
            imdbVotes = json_object["props"]["pageProps"]["aboveTheFoldData"]["ratingsSummary"]["voteCount"]
        except:
            imdbVotes = ""
            print(f"No votes")
        
        movie_data.append({"Title":title,
                           "Type":mtype,
                           "Release":release,
                           "Runtime":runtime,
                           "Genre":genre,
                           "Nominations":noms,
                           "Metascore":mscore,
                           "imdbRating":imdbRating,
                           "imdbVotes":imdbVotes,
                           "imdbID":movieid,
                           "Budget":imdbbudget,
                           "Gross":wwgross})
        time.sleep(1)
        
    except Exception as e:
        print(e)
        pass

#Export to csv
imdb_scrape = pd.DataFrame(movie_data)
imdb_scrape.to_csv("../output_data/imdb_all_movie_data.csv",index=False)
