==> ETL_notebooks/part4_data_cleanup.ipynb <==
#Dependencies
import pandas as pd

#Read the csv
raw_data = pd.read_csv("../output_data/imdb_all_movie_data.csv")
print(raw_data["Title"].count())
raw_data.head()

#Date format
true_dates = raw_data.loc[raw_data["Release"].str.contains("None")==False]
true_dates["Release"] = true_dates["Release"].astype("datetime64")

#Filters
filter_df = true_dates.loc[(true_dates["Release"]>="2018-01-01")&
                           (true_dates["Release"]<"2023-01-01")&
                           (true_dates["Type"]=="Movie")&
                           (true_dates["Runtime"]>=5400)]

#Count the remaining rows
filter_df["Title"].count()

#Read the csvs for each platform
netflix = pd.read_csv("resources/wikipedia_netflix_movies_2018_to_2022.csv")
amazon = pd.read_csv("resources/wikipedia_amazon_movies.csv")
disney = pd.read_csv("resources/wikipedia_disney_plus_original_movies.csv")
hulu = pd.read_csv("resources/wikipedia_hulu_movies.csv")

#Add platform columns
netflix["Platform"] = "netflix"
amazon["Platform"] = "amazon"
disney["Platform"] = "disney"
hulu["Platform"] = "hulu"

#Combine streaming platforms
streaming = pd.concat([netflix,amazon,disney,hulu])

#Merge
merge_df = pd.merge(filter_df,streaming,on="Title",how="left")

#Find duplicate titles
dupes = merge_df.loc[merge_df["Title"].duplicated()]["Title"].unique()

#Find indexes for duplicate titles
dupes_df = merge_df.loc[merge_df["Title"].isin(dupes),:]

#Drop all data for those titles
titles_df = merge_df.drop(dupes_df.index)

#Fill null platform types
titles_df["Platform"].fillna("others",inplace=True)

#Export to csv
titles_df.to_csv("../output_data/clean_data_for_analysis.csv",index=False)
