==> ETL_notebooks/part5_currency_and_language_imdb_scrape.ipynb <==
#Dependencies
import pandas as pd
import requests
import json
import time
import bs4 as BeautifulSoup

#Read the csv
clean_df = pd.read_csv("../output_data/clean_data_for_analysis.csv")

#Get currency and language from IMDb
HEADERS = {"User-Agent": "Chrome/111.0 (PC; Windows 10 Home)"}

movie_data = []

movieid_list = [movieid for movieid in clean_df["imdbID"]]

record = 0

for movieid in movieid_list:
    
    try:
        
        url = f"https://www.imdb.com/title/{movieid}/"
        response = requests.get(url, headers=HEADERS)
        soup = BeautifulSoup.BeautifulSoup(response.text,"html.parser")
        money = soup.find("script", id="__NEXT_DATA__")
        json_object = json.loads(money.text)
        
        record += 1
        
        title = json_object["props"]["pageProps"]["aboveTheFoldData"]["titleText"]["text"]
        
        print(f"Movie: {record} of {len(movieid_list)}: {title}")
        
        try:
            currencyb = json_object["props"]["pageProps"]["mainColumnData"]["productionBudget"]["budget"]["currency"]
        except:
            currencyb = ""
            print(f"No currencyb")

        try:
            currencyg = json_object["props"]["pageProps"]["mainColumnData"]["worldwideGross"]["total"]["currency"]
        except:
            currencyg = ""
            print(f"No currencyg")
        
        try:
            language = json_object["props"]["pageProps"]["mainColumnData"]["spokenLanguages"]["spokenLanguages"][0]["text"]
        except:
            language = ""
            print(f"No language")
        
        movie_data.append({"imdbID":movieid,
                           "CurrencyBudget":currencyb,
                           "CurrencyGross":currencyg,
                           "Language":language})
        time.sleep(1)
        
    except Exception as e:
        print(e)
        pass

#Export to csv
imdb_scrape = pd.DataFrame(movie_data)
imdb_scrape.to_csv("../output_data/currency_language.csv",index=False)
