==> ETL_notebooks/part6_merge_currency_language.ipynb <==
#Dependencies
import pandas as pd

#Read csvs into DataFrames
main_df = pd.read_csv("../output_data/clean_data_for_analysis.csv")
curr_lang_df = pd.read_csv("currency_language.csv")

#Merge DataFrames
clean_df = pd.merge(main_df,curr_lang_df,on="imdbID",how="left")
print(clean_df["Title"].nunique())

#Export to csv
clean_df.to_csv("../output_data/clean_with_currency_language.csv",index=False)
